refactor: restructure dashboard components and update layout for improved organization
