+++ src/app/(dashboard)/_components/dashboard-header.tsx
@@ -1,7 +1,7 @@
 "use client";
-import { SidebarToggleButton } from "./admin-sidebar";
+import { SidebarToggleButton } from "./dashboard-sidebar";
 
-export default function Header() {
+export default function DashboardHeader() {
   return (
     <div className="-mx-[var(--_sidebar-spacing)] px-[var(--_sidebar-spacing)] py-[var(--_sidebar-spacing)] sticky top-0 flex items-center gap-[var(--_sidebar-spacing)] bg-sidebar border-b border-border h-[var(--_sidebar-header-height)] shadow-2xl">
       <SidebarToggleButton /> <Separator />
