+++ src/app/(dashboard)/_components/index.ts
@@ -1,5 +1,5 @@
-export { default as Sidebar } from "./admin-sidebar";
+export { default as DashboardSidebar } from "./dashboard-sidebar";
 export { default as SidebarProvider } from "./sidebar-provider";
-export { default as AdminContainer } from "./admin-container";
-export { default as AdminHeader } from "./admin-header";
+export { default as DashboardLayout } from "./dashboard-layout";
+export { default as DashboardHeader } from "./dashboard-header";
 export { useSidebar } from "./sidebar-provider";
